Move Table files to /src/Table
